update page to use better html / css
seriously, it is still pretty bad. Like, why did it use position absolute? The layout it had was pretty easy to make without position absolute. Plus, many things were in illogical places, like the preview was after the editor, which makes no sense, because the preview is before the editor. Other weird things were, the export options were being added at the end of the body instead of before the preview.

I may have to go in and redo the entire css to make it much much better.