snapblocks rebrand
I feel like I've started to add enough to warrant completely renaming it to snapblocks, as well as resetting the version number to 1.0.0. this is only a package name change, the actual code still uses scratchblocks as the main name in the code. This is so it's easy to transition from scratchblocks to snapblocks. I don't know if I'll change this though.