add multiline block support
in blocks, you need to type \n (yeah, I know, snap uses $nl), and in string inputs, you can just hit enter (\n still works). It took me a long time to get this to work with c-shaped blocks.

Oh, and this is currently only for the snap style, but I'll try to add it to scratch 2 and 3.