add icon scale and color
I actually love this so much, because now I can scale icons to whatever size I want (this is actually used for the loop arrow with the :: loop override). Custom blocks don't need to have the same icon scale / color because I think that would be a little too strict. Although there is one minor problem, this new feature also scales the drop-shadow, which might not be the best.