add / change icons, add all changes to scratch 2, and more
I added some new icons, as well as copying a bunch of the changes I made in the snap style to the scratch 2 style. I'm going to work on moving them to scratch 3 next. I also changed the minimum input width for the snap style.

Should I add the snap visual space dot? I'm not sure.