add new block detection system, and more
This new block detection system is a dynamic one, using abstract block specs. The amazing thing about this, is that it's able to detect variadic inputs. Seriously, I spent so long thinking of ideas to make variadic inputs in an elegant way, but this is the one that I was like, wait, that might actually work.

Now, this new block detection system is still a work in progress, as I have not added proper translation support (and I basically removed the function to translate the blocks), and it doesn't factor in any of the disambiguations, but it still works.

You can test this out with the snap if block, yes the variadic one.

```
if <> {

} else if <> {

} else if <> {

} @delInput @addInput
```

Alright, nobodies here, AAAAAAAAAAAAAAAAAAAAAHHHHHHHHHHHHHHHHHHHHHHHHH!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! I FINALLY MADE VARIADIC INPUTS THAT ACTUALLY WORK!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!