add remaining snap blocks
I also had to add a few more disambiguations, and I had to change the "of" disambiguation to be a little weird, because it's a 4-way collision. This is similar to the when block, but that one was a bit easier to deal with (I just had to add a check to the snap when block). The length of blocks were a bit weird, since I decided to support the @list icon with a def, and my code apparently doesn't bother to check other defs, if it found specs that are equal. This is just due to my coding, so I hope this'll be fixed in a future version. For now, the specialCase function is working.