fix up tests so they all pass my new formatting
I also made sure to make the tests that should pass, pass.

I did have to remove the ([e^ v] of (9)) test, because that is now a valid math function (snap does not have a space in it).

Later, I'm planning on adding some more tests, specifically for the new things I'm adding.