add category colors into code
Yay! I managed to put snap's Color class into snapblocks, and use that. This is great, because I no longer have to rely on css classes, and the zebra coloring is exact (the exact same lighter function as snap). In fact, since it's in the code, I may be able to use these colors elsewhere, such as if I ever create syntax highlighting.