add loop arrows on any c-slot
If there's an @loop icon at the end of a line right after a c-slot, it will be treated as a loop arrow for the c-slot. This allows you to set a loop arrow on any c-slot.

Oh, I also added an alias for @loopArrow, you can now just type @loop.