set scale to 1 = size = 10
I was originally doing 1 = size = 12, because 12 was the default size of icons in snap, but I just discovered that if you set the scale on an icon, the size is set to 10 (1 = 10), so I am making this change to account for that. Currently I'm just doing the quick and dirty, multiply the scale by 10/12, but hopefully in the next update, I'll (try) to redo all the icons so they're made for a size of 10.