change flag icon and add greenflag
I added a flag icon, which is the same as the greenFlag icon, but without color or scaling (this will also allow me to add the new blitz icon in snap 10). Basically just what snap 10 does. I also added an alias for greenflag, specifically because that's what snap 10 uses for the scaled and colored flag icon.