fix upvar zebra color
I found out that upvars are supposed to have the same color as their parent, including the zebra coloring, but there's a bug where it doesn't update automatically. If you put it in a custom block, close the custom block editor, then open it, the upvar color will be the zebra color.